Removed test for invalid content-type header
This test is no longer valid because CakePHP automatically sets the content-type of the outgoing
message based on two factors: whether or not it has attachments, and whether its format is text,
HTML, or both. In all of these cases, a valid content-type is set, so this test is obsolete.